feat: communicate with i2c oled
u8x8 has been added as a library to control a 0.96" I2C OLED screen. I2C
was implemented as well in master transmitter mode to be able to
communicate with said display.

Macro magic was also used to improve the coding experience and not have
to write weird keywords with '@'.