feat:  implement led1642
Communication with the led controller onboard has been enabled. It
successfully sets the PWM counter in 12 bit mode and sets brightness
levels.

I2C has been disabled because really nothing is being done with the
screen and the code is unpolished, it often times seems to crash the
MCU resulting in a corrupted screen and unresponsive system.